zreader.c: initial revision, replaces librpmio
So rpmio is a mess.  They got it wrong, back then and now.  It strives
to provide a "stdio clone" API, as if it were something good.  And
stdio is a complex beast, and a mess, too.  But at least stdio is a mess
for a reason.  In stdio, file handles can be readable, writable, OR BOTH,
and on top of that, seekable.  That's the reason for the unified FILE
handle as we know it.  As opposed to this, compressed streams can be
either readable or writable, but not both; readable compressed streams
are not easily seekable; writable compressed streams are never seekable.
As a result, rpmio provides a leaky abstraction driven by a false analogy.

There's no point then in trying to put all IO under the unified file
handle.  A better approach could be to provide separate handles for
sequential reading and sequential writing of compressed streams.
This is what this new zreader.c library is all about.  Only gzip
and lzma/xz decompression is implemented, but that covers just about
every package I need to process.  Support for bzip2 and zstd can be
added later.