header.c: replaces librpm:rpmReadPackageFile and headerGet
Now, librpm is a mess too.  Loading a header is by no means lightweight.
When a header is loaded, librpm somehow "sorts" it (which you immediately
know if you take a look at the callgrind_annotate output - msort_with_tmp
will be at the very top).  So they got it wrong, too, back then and now.

Anyway, this new library implements lightweight parsing: only the tags
which are needed for the rpmcpio library will be searched, in a single
pass.  For now, this is just Dirindexes, which provides fileCount.
PayloadCompressor is hardcoded to "lzma", so this is an early
barely working version.