header.c: implemented header_find()
I was worried that looking up file info for each cpio entry could make
things slower, and devised a few optimization.  The main optimization
takes advantage of the fact that cpio entries are mostly sorted (by
filename), so only one comparison would suffice in most cases.  (The
comparison is integrated in the bsearch loop, so nothing is wasted.)
The second optimization is that, when repeatedly matching against
the same dirname in fi->dn, the dir part of comparison can be cached.

Wow, it's so fast now, even with gzip payloads,
I have to pass --threshold=100 to callgrind_annotate.

Apparently I have developed quite a propensity to optimizing string
routines (among other things).  What a great artist is dying in me!
Of a burnt theater...