header.c: split file info into ffi[] and ffx[]
Extended / excluded cpio entries do not provide metadata at all.
Not even the size of the attached data, nor filename.  All the
information has to be obtained from the header.  What's even more funny
is that the header does not store nlink (link counts).  Those have to be
restored upon header loading by counting flies with the same ino+dev.
So the answer to the question "what comes after 07070X?" is very far
form obvious (hardlink counting pass is not implemented just yet).

Anyway, I prefer to split file info into to parts: the basic part,
or ffi[], which is required in all cases (e.g. to provide fflags).
And the extended part, or ffx[], which provides the information
missing in cpio extended entries.  For packages with no large files,
h->ffx will be set to NULL.

I'm also trying out a new approach to uid/guid.  In cpio, these IDs
are numeric (from the build environment); in rpm header, they are stored
as strings.  There is no direct correspondence, and rpm does not ensure
that all the IDs exist before installing a package.  However, it is safe
to assume that "root" will be mapped to 0.  The distinction between root
and non-root is just about the only reliable fact about the IDs.

I also relax some checks.  The general principle is that we check only
the tags which we are going to load.  Thus if ffx[] entries are not
loaded, some checks can be omitted.  We also make check against
ambiguity, but only very basic ones.  For example, if OLDFILENAMES are
set, we only check against BASENAMES, but do not care about DIRNAMES
or DIRINDEXES.