header.c: removed support for uid / gid
IDs are numeric in the archive and strings in the header.  I can't make
any sense out of out.  Should the entry then expose numbers or strings?
Besides, there is no reliable correspondence between the numbers and the
strings.

Why do I even want this?  Well, it would be nice, I thought, to show
IDs along with sizes and mtimes, much like ls(1).  But for that, they
already have headerSprintf and 'rpm -ql'.  And rpmcpio is primarily
for processing file data.

"If a problem is not completely understood, it is probably best
to provide no solution at all."