header.c: simplify malloc, tighter array processing
The code is way too verbose, and the single malloc calculation is too
error-prone: I wasn't able to convince myself that the code always works
correctly.  Therefore, temporary space is now allocated (and freed)
separately.

This also opens up the opportunity for more efficient array loading.
Although reada() is inlined, calling it for every array element results
in loops that are not as tight.  This was done because values need to be
scattered into ffi[] fields.  Now the values are read with a single
reada call, and then scattered in a tight loop.