header.c: stricter offset validation in the merge loop
The code should be reliable, i.e. not easily broken by malformed data.
On the other hand, the code should remain general.  For example, should
it accept a header with hdr.il==0 or hdr.dl==0?  Well, if nothing
breaks, there is not enough reason to reject it.  It should accept then.
(This is only one consideration.  The code is still a bit too
complicated.  The alternative is to validate in a separate loop.)