header.c: remove fildevices
Files should be identifiable by inode only (otherwise this complicates
the user-visible part of the API).  Modern rpm-build renumbers inodes
anyway.  With older rpm-build, inodes are unique as long as %buildroot
is set properly during the build.  The only rpms in danger are those
built in the host system spreading the files across a few mount points.
This is a thing of the very distant past.