header.c: header_find returns gracefully on h->fileCount == 0
"Can't call me if h->fileCount == 0" is too strong a precondition.
It's far from intuitive, and it seems that just about the only way
to meet the condition is that the caller must explicitly check for
h->fileCount == 0 before calling header_find().  So there's a better
way to do it in header_find(): the check is performed only when
++h->prevFound goes out of bounds.